Add deprecation notice for lmc(4)
We intend to remove support before FreeBSD 12 is branched.  These are
available only as 32-bit PCI devices.  The driver has an ambiguous
license and I have not been successful in contacting the driver's author
in order to address this.

The planned deprecation has been announced on -current and -stable; if
we receive feedback that the driver is still useful and we are able to
resolve the license issue this deprecation notice can be reverted.

Reviewed by:	bapt, brooks, imp, rgrimes
MFC after:	2 weeks
Relnotes:	Yes
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D15182


git-svn-id: svn+ssh://svn.freebsd.org/base/head@332966 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f