Reduce overhead of ktrace checks in the common case.
KTRPOINT() checks both if we are tracing _and_ if we are recursing within
ktrace. The second condition is only ever executed if ktrace is actually
enabled. This change moves the check out of the hot path in to the functions
themselves.

Discussed with mjg@

Reported by:	mjg@
Approved by:	sbruno@


git-svn-id: svn+ssh://svn.freebsd.org/base/head@333390 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f