aw_mmc: Do not fully init the controller in attach
Only do a reset of the controller at attach and init it at power_up.
We use to enable some interrupts in reset, only enable the interrupts
we are interested in when doing a request.
While here remove the regulators handling in power_on as it is very wrong
and will be dealt with in another commit.

Tested on: A31, A64


git-svn-id: svn+ssh://svn.freebsd.org/base/head@333555 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f