sx: fixup a braino in r334024
If a thread waiting on sx dropped Giant it would not be properly
reacquired on exit from the routine, later resulting in panics
indicating Giant is not held (when it should be).

The bug was not present in the original patch sent to pho, I wittingly
added it just prior to the commit and only smoke-tested it.

Reported by:	pho


git-svn-id: svn+ssh://svn.freebsd.org/base/head@334048 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f