Add the "-p" and "-m" options to nfsd.c for the pNFS service.
The "-p" option specifies that the nfsd should run a pNFS service instead
of a regular NFS service. The "-m" option is only meaningful when used with
"-p" to specify that mirroring on the DSs should be done and on how many of
them.
This change requires the kernel changes committed as r334930.
The man page update will be committed as a separate commit soon.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@335172 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f