This patch is actually all client side changes and does not affect the server.
Get rid of newnfs_canceldspreqs(), since it causes problems when there are
multiple DSs on the same machine and they share a TCP connection.

Add a wakeup() plus fill in error information for Layoutreturns, so that
a layoutreturn with the error information gets sent to the MDS quickly.

Make a slight change so that a layoutreturn with error information is
returned to the MDS for all errors except EACCES. Also, pass the error
back to the MDS instead of just EIO so that the MDS can decide what to
do w.r.t. it.