Add a new "-m" option to pnfsdsfile(8) to prepare a file for mirroring.
When pnfsdscopymr(8) is used to create a mirror of a file on a mirrored
pNFS service, it expects to find an entry in the extended attribute for
IP address 0.0.0.0.
This patch adds a "-m" option which can be used to create these entrie(s).
It also tightens up the checks for use of incompatible command line options.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@335845 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f