Modify the pNFS server kernel code so that a "force" option can be implemented for pnfsdskill(8).
pnfsdskill(8) will fail if there isn't a mirror for the DS that is being
disabled. However, if the DS has failed, it may be necessary to do a
pnfsdskill on the DS even when there is no valid mirror, so that the
nfsd threads can be killed.
This patch modifies the kernel to add a PNFSDOP_FORCEDELDS option that
pnfsdkill(8) can use to force the disable.