Fix the NFSv4.1 pNFS client for the case of mirrored DSs on separate machines.
My test environment has all the DSs colocated on the same machine, so I did
not spot this bug until recently. The client was using the TCP connection
for the first DS when doing writes/commits to subsequent mirrored DSs.
This worked fine for my test setup, but would have failed when the DSs are
on separate machines.
This patch fixes the code so that it uses the correct TCP connection for
the writes/commits to each of the mirrored DSs.