Add a timeout to the msleep() waiting for a socket connection for the pNFS client.
Without this patch, the client TCP connection code can wait a long time
for a connection attempt to fail when a DS is not operational.
This patch adds a timeout for this msleep() if the CLSET_TIMEOUT has been
set for the connection. This is only set for TCP connections to DSs.