aw_thermal: Add nvmem and H5 support
Now that aw_sid expose nvmem interface, use that to read the calibration
data.
Add support for H5 SoC.
Fix the bindings, we used to have non-upstreamed bindings. Switch to the
one that have been sent upstream. They are not stable yet, so we switch
from custom, wrong, bindings to correct, proposed bindings


git-svn-id: svn+ssh://svn.freebsd.org/base/head@337372 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f