libbe(3): Fix be_import to delete temp snapshot
Deleting the temp snapshot isn't immediately possible because it's the
origin of the newly imported boot environment. However, this is trivially
solved by opening the new boot environment and promoting it. The roles are
now reversed and the temp snapshot/dataset may be completely destroyed.

Remove the BUGS from libbe(3) and bectl(8).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@337697 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f