lualoader: Fix override of module_path on loader prompt
Earlier changes setup a config.module_path variable that was populated upon
reading of loader.conf(5) and used for restoring module_path to pristine
condition if multiple kernels are attempted. This broke the ability to
override module_path at the loader prompt in case of emergency.

Approved by:	re (rgrimes)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@338309 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f