Allow `capsicum-test` to be executed from an absolute path by stashing the execution directory
This change stashes the executing directory for `capsicum-test` then uses it to
determine where `mini-me*` lives, allowing the script to be executed from an
absolute path.

This change required refactoring how the fexecve tests were executed, as the
path for `mini-me*` could no longer be determined at compile-time, but needs
to be determined at runtime.

As such, switch from FORK_TEST to FORK_TEST_F (with an appropriate class) and
FORK_TEST_ON to FORK_F with proper setup/teardown fixtures for cleaning up the
temporary script.