Merge tests/sys/capsicum-test into tests/sys/capsicum
One of the stopgap issues with compiling capsicum-test in tests/sys/capsicum
was that WARNS.capsicum-test wasn't overriding WARNS, resulting in compilation
failures when the two were mixed.

Now that D19755 is being fixes, WARNS.capsicum-test is able to properly
override WARNS, making it possible for the two to coexist.