Fix an item and run more tests
First off, the BINMODE for mini-me.setuid should be 04555, not 04444. The
latter mode is not executable, so mini-me.setuid would do the wrong thing by
failing to run. There's probably a capsicum-test bug buried in here.

Second off, some of the *[Fe|E]xecve tests are skipped if run as root. In order
to ensure these tests aren't skipped, run them as an unprivileged user.