Improve googletest integration with kyua
This change modifies googletest.test.mk to use experimental googletest support
in kyua, available as part of the `add-googletest-support` feature branch in
https://github.com/ngie-eign/kyua. The experimental support provides test suite
and test case level execution and result reporting, instead of requiring
googletests to be run as the plain tests.

As such, make the default GTESTS test engine interface googletest, providing
a backwards compatibility knob, GTESTS_USE_PLAIN_TEST_INTERFACE, so
end-users with older versions of kyua can run the googletest tests.

While here, provide a per-PROG test interface hook for GTESTS. In short,
not all of the googlemock and googletest provided tests conform to the
format described in the googletest docs. Thus, they are better run as plain
tests.