Build libclang_rt.profile as a cross-tools dependency
Long story short, make tinderbox was failing because it was looking for
libclang_rt.profile under a different path from there make libraries installs
it. Add a new MK knob to control building `lib/libclang_rt/profile` in
buildworld: MK_COVERAGE_SUPPORT. MK_COVERAGE_SUPPORT is to
`lib/libclang_rt/profile`, as MK_TESTS_SUPPORT is to lib/atf/...: it serves to
separate out building infrastructure from the resulting code.

This commit is untested, but I'm hopeful that it's a move in the right
direction.