Disable MK_COVERAGE* when cross-building
Unfortunately, make tinderbox is building a copy of libclang_rt.profile*.a for
the machine architecture, not the target architecture, while looking for the
library under the cross-toolchain directory, with the current set of proposed
changes. As such, disable MK_COVERAGE and MK_COVERAGE_SUPPORT until I work out
why the library isn't being built and installed as expected.

Add a comment to note the issue.