Take a more relaxed approach to CXXSTD
Only inject `-std=${CXXSTD}` if the value of CXXSTD is not empty.

This satisfies the backwards compatibility of bsd.sys.mk, pre-r345708, and the
aim of CXXSTD in r345708, without having to hardcode nop values and build
unnecessary intelligence of supported C++11 standards in bsd.sys.mk for each
compiler.